src/requirements.txt
ansible
AudioConverter
beautifulsoup4
bokeh
boto3
catboost
cffi
chainer
click
config
# cupy
Cython
decorator
django
dlib
dtreeviz
fastapi
fasttext
ffmpeg 
ffmpeg-python
flask
flask-marshmallow
flask-ngrok
gensim
ggplot
git+https://github.com/tensorflow/docs
google-api-python-client
google-auth
google-cloud-bigquery
google-cloud-storage
google-cloud-vision
graphviz 
gym
h5py
httpie
imageio
imageio-ffmpeg
imbalanced-learn
iniconfig
ipykernel
ipympl
ipython
ipywidgets
janome
joblib 
JSAnimation
jupyter
jupyterlab
japanize-matplotlib
keras-tuner
lameenc
librosa
lightgbm
many-stop-words
matplotlib
mecab-python3
mlflow
moviepy
music21
mysqlclient
mysql-connector-python
neo4j
nest-asyncio
networkx
nltk 
numba
numpy
opencv-python
opencv-python-headless
openpyxl
optuna
pandas
pandas-profiling
pdfminer
pdfminer3
pdfminer.six
pillow
pip
plotly
psycopg2
PuLP
pyautogui
pycaret
pydotplus
pydub
PyExcelerate
pygal
pygame
pyinstaller
pylint
pymc3
pyngrok
PySimpleGUI
pytesseract
python-Levenshtein
python-pptx
python-xlib
pyvirtualdisplay
pyyaml
resampy 
requests
scikit-learn
scipy 
scrapy
seaborn 
selenium
seqeval 
setuptools 
shap
shapely
SoundFile
SQLAlchemy
sympy
tensorboard 
tensorboardX 
tensorflow-addons
# tensorflow-datasets
tensorflow==2.5.0
tensorflow-hub
tensorflow-io
# tensorflow-text
texthero
tfds-nightly
# tf-models-official
# tf-nightly
toml
torch
torchaudio 
torchvision
tqdm
transformers 
twine
uvicorn
webdriver_manager
wget
wikiextractor
xgboost
xlrd==1.2.0
XlsxWriter
youtube-dl